Move remaining structs, the only ones left are unused (from other .dlls)
