+++ source/D2Common/include/D2Composit.h
@@ -41,6 +41,12 @@ enum D2WeaponClasses
 	WEAPONCLASS_HT2
 };
 
+struct D2CompositStrc
+{
+	int32_t nWeaponClassCode;					//0x00
+	int32_t nWeaponClassId;						//0x04
+};
+
 //D2Common.0x6FD466C0 (#10884)
 void __stdcall D2Common_10884_COMPOSIT_unk(D2UnitStrc* pUnit, int nClass, int nMode, int nUnitType, D2InventoryStrc* pInventory, char* szPath, int* pWeaponClassCode, BOOL bAddPathPrefix, int a9);
 //D2Common.0x6FD46BC0 (#10885)
