+++ source/D2Common/include/D2Text.h
@@ -18,6 +18,31 @@ struct D2MessageListStrc		//sizeof 0x22
 	D2MessageStrc pMessages[8];	//0x02
 };
 
+struct D2TextHeaderStrc
+{
+	void* pMemPool;				//0x00
+	uint16_t nCount;			//0x04
+	uint16_t pad0x06;			//0x06
+	D2TextNodeStrc* pNode;		//0x08
+};
+
+struct D2TextMessageStrc
+{
+	wchar_t* pStringLines[6];	//0x00
+	uint16_t nLines;			//0x18
+	int32_t nColor;				//0x1A
+	uint32_t nEndTick;			//0x1E
+	D2TextMessageStrc* pNext;	//0x22
+};
+
+struct D2TextNodeStrc
+{
+	int16_t nStringId;			//0x00
+	uint16_t pad0x02;			//0x02
+	int32_t nMenu;				//0x04
+	D2TextNodeStrc* pNext;		//0x08
+};
+
 //D2Common.0x6FDC36E0 (#10901)
 D2TextHeaderStrc* __stdcall TEXT_AllocTextHeader(void* pMemPool);
 //D2Common.0x6FDC3710 (#10902)
