+++ source/D2Common/include/Drlg/D2DrlgDrlgAnim.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "CommonDefinitions.h"
+#include "D2DrlgDrlg.h"
 #include "D2DrlgDrlgGrid.h"
 
 //D2Common.0x6FD75480
