+++ source/D2Common/include/Drlg/D2DrlgPreset.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "CommonDefinitions.h"
+#include "D2DrlgDrlg.h"
 #include "D2DrlgDrlgGrid.h"
 
 struct D2UnkFileStrc
@@ -9,6 +10,18 @@ struct D2UnkFileStrc
 	long field_104;						//0x104
 	D2DrlgFileStrc* pFile;				//0x108
 	D2UnkFileStrc* pNext;				//0x10C
+};
+
+struct D2PresetUnitStrc
+{
+	int32_t nUnitType;						//0x00
+	int32_t nIndex;							//0x04 see D2Common.#11278
+	int32_t nMode;							//0x08
+	int32_t nXpos;							//0x0C
+	int32_t nYpos;							//0x10
+	BOOL bSpawned;							//0x14 
+	D2MapAIStrc* pMapAI;					//0x18
+	D2PresetUnitStrc* pNext;				//0x1C
 };
 
 struct D2DrlgMapStrc
