+++ source/D2Common/include/Path/Path.h
@@ -11,6 +11,120 @@ enum D2PathTypes
 	PATHTYPE_KNOCKBACK_CLIENT = 11,
 };
 
+struct D2PathPointStrc
+{
+	uint16_t X;
+	uint16_t Y;
+};
+
+struct D2DynamicPathStrc
+{
+	union
+	{
+		struct
+		{
+			uint16_t wOffsetX;					//0x00
+			uint16_t wPosX;						//0x02
+			uint16_t wOffsetY;					//0x04
+			uint16_t wPosY;						//0x06
+		};
+
+		struct
+		{
+			uint32_t dwPrecisionX;				//0x00
+			uint32_t dwPrecisionY;				//0x04
+		};
+	};
+	uint32_t dwTargetX;						//0x08
+	uint32_t dwTargetY;						//0x0C
+	uint16_t xSP1;								//0x10
+	uint16_t ySP1;								//0x12
+	uint16_t xSP2;								//0x14
+	uint16_t ySP2;								//0x16
+	uint16_t xSP3;								//0x18
+	uint16_t ySP3;								//0x1A
+	D2RoomStrc* pRoom;						//0x1C
+	D2RoomStrc* pRoomNext;					//0x20
+	int32_t unk0x24;							//0x24
+	int32_t dwPathPoints;						//0x28
+	void* unk0x2C;							//0x2C
+	D2UnitStrc* pUnit;						//0x30
+	uint32_t dwFlags;							//0x34
+	uint32_t unk0x38;							//0x38
+	uint32_t dwPathType;						//0x3C
+	uint32_t dwPrevPathType;					//0x40
+	uint32_t dwUnitSize;						//0x44
+	uint32_t dwCollisionPattern;				//0x48
+	uint32_t dwCollisionType;					//0x4C
+	uint32_t unk0x50;							//0x50
+	uint16_t unk0x54;							//0x54
+	uint16_t unk0x56;							//0x56
+	D2UnitStrc* pTargetUnit;				//0x58
+	uint32_t dwTargetType;						//0x5C
+	uint32_t dwTargetId;						//0x60
+	uint8_t nDirection;						//0x64
+	uint8_t nNewDirection;						//0x65
+	uint8_t nDiffDirection;					//0x66
+	uint8_t unk0x67;							//0x67
+	uint8_t unk0x68[10];						//0x68
+	int32_t unk0x72;							//0x72
+	int32_t unk0x76;							//0x76
+	char unk0x7A[2];						//0x7A
+	uint32_t dwVelocity;						//0x7C
+	uint32_t unk0x80;							//0x80
+	uint32_t dwVelocity2;						//0x84
+	uint32_t unk0x88[2];						//0x88
+	uint8_t nDist;								//0x90
+	uint8_t nDistMax;							//0x91
+	uint8_t unk0x92;							//0x92
+	uint8_t nStepNum;							//0x93
+	uint8_t nDistance;							//0x94
+	char unk0x95[3];						//0x95
+	//uint32_t dwPathOffset;						//0x98
+	D2PathPointStrc PathPoints[7];			//0x98
+	uint32_t unk0xB8[72];						//0xB8
+	int32_t unk0x1D4;							//0x1D4
+	D2PathPointStrc unk0x1D8[7];			//0x1D8
+	char unk0x1DC[12];						//0x1DC
+};
+
+struct D2PathInfoStrc
+{
+	D2PathPointStrc pStartCoord;			//0x00
+	D2PathPointStrc field_4;				//0x04
+	D2RoomStrc* pRoom;						//0x08
+	D2RoomStrc* field_C;					//0x0C
+	int32_t field_10;							//0x10
+	uint8_t field_14;							//0x14
+	uint8_t field_15;							//0x15
+	uint16_t field_16;							//0x16
+	int32_t nDistMax;							//0x18
+	uint8_t field_1C;							//0x1C
+	uint8_t field_1D;							//0x1D
+	uint16_t field_1E;							//0x1E
+	int32_t nPathType;							//0x20
+	int32_t nUnitSize;							//0x24
+	int32_t nCollisionPattern;					//0x28
+	int32_t nCollisionType;						//0x2C
+	union
+	{
+		D2DynamicPathStrc* pDynamicPath;	//0x30 - not sure yet
+		D2PathInfoStrc* pNext;				//0x30
+	};
+};
+
+struct D2StaticPathStrc
+{
+	D2RoomStrc* pRoom;						//0x00
+	int32_t nTargetX;							//0x04
+	int32_t nTargetY;							//0x08
+	int32_t nXPos;								//0x0C
+	int32_t nYPos;								//0x10
+	uint32_t unk0x14[2];						//0x14
+	uint8_t nDirection;						//0x1C
+	uint8_t unk0x1D[3];						//0x1D
+};
+
 //D2Common.0x6FDA8220
 void __fastcall sub_6FDA8220(D2DynamicPathStrc* pDynamicPath);
 //D2Common.0x6FDA82A0 (#10141)
