+++ source/D2Common/include/Units/MissileStream.h
@@ -2,6 +2,12 @@
 
 #include "CommonDefinitions.h"
 
+struct D2MissileStreamStrc
+{
+	int32_t* unk0x00;							//0x00
+	int32_t unk0x04;							//0x04
+};
+
 
 //D2Common.0x6FDBC230 (#11213)
 void __stdcall MISSTREAM_AllocMissileStream(D2UnitStrc* pMissile);
