+++ source/D2Common/src/Drlg/DrlgDrlgLogic.cpp
@@ -1,6 +1,7 @@
 #include <D2Seed.h>
-#include "Drlg/D2DrlgDrlgLogic.h"
 
+#include "Drlg/D2DrlgDrlg.h"
+#include "Drlg/D2DrlgDrlgLogic.h"
 #include "Drlg/D2DrlgDrlgAnim.h"
 #include "Drlg/D2DrlgDrlgGrid.h"
 #include "Drlg/D2DrlgDrlgRoom.h"
