+++ source/D2CommonDefinitions/include/D2Constants.h
@@ -86,31 +86,6 @@ enum D2C_ClientStates
 	CLTSTATE_CONNECTED = 4
 };
 
-enum D2C_RosterInfoFlags
-{
-	ROSTERINFOFLAG_LOOT = 0x01,
-	ROSTERINFOFLAG_IGNORE = 0x02,
-	ROSTERINFOFLAG_SQUELCH = 0x04,
-	ROSTERINFOFLAG_HOSTILE = 0x08,
-};
-
-enum D2C_RosterControlFlags
-{
-	ROSTERCONTROLFLAG_INVITE = 0,
-	ROSTERCONTROLFLAG_INPARTY = 1,
-	ROSTERCONTROLFLAG_ACCEPT = 2,
-	ROSTERCONTROLFLAG_LEAVE = 3,
-	ROSTERCONTROLFLAG_CANCEL = 4,
-};
-
-enum D2C_RosterButtonTypes
-{
-	ROSTERBUTTONTYPE_HOSTILE = 1,
-	ROSTERBUTTONTYPE_LOOT = 2,
-	ROSTERBUTTONTYPE_IGNORE = 3,
-	ROSTERBUTTONTYPE_SQUELCH = 4,
-};
-
 enum D2SaveCharType
 {
 	CHARTYPE_HARDCORE = 0x24,
