+++ source/D2CommonDefinitions/include/D2Structs.h
@@ -261,12 +261,6 @@ struct D2MercSaveDataStrc
 	uint32_t dwExperience;						//0x0C		//0xBB
 };
 
-struct D2PathPointStrc
-{
-	uint16_t X;
-	uint16_t Y;
-};
-
 struct D2SaveSkillKeyStrc
 {
 	uint16_t nSkill;							//0x00
@@ -633,11 +627,6 @@ struct D2CombatStrc
 	D2CombatStrc* pNext;					//0x18
 };
 
-struct D2CompositStrc
-{
-	int32_t nWeaponClassCode;					//0x00
-	int32_t nWeaponClassId;						//0x04
-};
 
 struct D2CorpseStrc
 {
@@ -722,485 +711,6 @@ struct D2DamageStrc
 //	D2DrlgMiscDataStrc* unk0xBC;			//0xBC
 //	uint32_t unk0xC0[11];						//0xC0
 //};
-
-struct D2DrlgCoordStrc
-{
-	int32_t nPosX;								//0x00
-	int32_t nPosY;								//0x04
-	int32_t nWidth;								//0x08
-	int32_t nHeight;							//0x0C
-};
-
-struct D2DrlgCoordsStrc
-{
-	int32_t dwSubtilesLeft;						//0x00
-	int32_t dwSubtilesTop;						//0x04
-	int32_t dwSubtilesWidth;					//0x08
-	int32_t dwSubtilesHeight;					//0x0C
-	int32_t dwTilesLeft;						//0x10
-	int32_t dwTilesTop;							//0x14
-	int32_t dwTilesWidth;						//0x18
-	int32_t dwTilesHeight;						//0x1C
-};
-
-struct D2RoomExStrc
-{
-	D2DrlgLevelStrc* pLevel;				//0x00
-	union
-	{
-		struct
-		{
-			int32_t nTileXPos;					//0x04
-			int32_t nTileYPos;					//0x08
-			int32_t nTileWidth;					//0x0C
-			int32_t nTileHeight;				//0x10
-		};
-		D2DrlgCoordStrc pDrlgCoord;			//0x04
-	};
-	uint32_t dwFlags;							//0x14
-	uint32_t dwOtherFlags;						//0x18
-	int32_t nType;								//0x1C
-	union
-	{
-		D2DrlgPresetRoomStrc* pMaze;		//0x20
-		D2DrlgOutdoorRoomStrc* pOutdoor;	//0x20
-	};
-	uint32_t dwDT1Mask;						//0x24 - tile caching mask, used to init below @ D2Common.0x4A380
-	D2TileLibraryHashStrc* pTiles[32];		//0x28
-	D2DrlgTileGridStrc* pTileGrid;			//0xA8
-											//uint32_t fRoomStatus;						//0xAC - actually a byte, the padding up to +B8 is related
-	uint8_t fRoomStatus;						//0xAC
-	uint8_t unk0xAD;							//0xAD
-	uint16_t unk0xAE[5];						//0xAE
-	D2RoomExStrc* pStatusNext;				//0xB8
-	D2RoomExStrc* pStatusPrev;				//0xBC
-	D2RoomExStrc** ppRoomsNear;				//0xC0 // names pptVisibleRooms in the original game
-	int32_t nRoomsNear;							//0xC4
-	D2RoomTileStrc* pRoomTiles;				//0xC8
-	D2PresetUnitStrc* pPresetUnits;			//0xCC
-	D2DrlgOrthStrc* pDrlgOrth;				//0xD0
-	D2SeedStrc pSeed;						//0xD4
-	uint32_t dwInitSeed;						//0xDC
-	D2DrlgCoordListStrc* pCoordList;		//0xE0 // named ptLogicalRoomInfo in the original game
-	D2RoomStrc* pRoom;						//0xE4
-	D2RoomExStrc* pRoomExNext;				//0xE8
-};
-
-typedef int32_t(__stdcall* ROOMCALLBACKFN)(D2RoomStrc*, void*);
-
-struct D2RoomStrc
-{
-	union
-	{
-		struct
-		{
-			int32_t nSubtileX;					//0x00
-			int32_t nSubtileY;					//0x04
-			int32_t nSubtileWidth;				//0x08
-			int32_t nSubtileHeight;				//0x0C
-			int32_t nTileXPos;					//0x10
-			int32_t nTileYPos;					//0x14
-			int32_t nTileWidth;					//0x18
-			int32_t nTileHeight;				//0x1C
-		};
-		D2DrlgCoordsStrc pCoords;			//0x00
-	};
-	D2DrlgRoomTilesStrc* pRoomTiles;		//0x20
-	D2RoomStrc** ppRoomList;				//0x24
-	int32_t nNumRooms;							//0x28
-	D2UnitStrc* pUnitFirst;					//0x2C
-	D2UnitStrc* pUnitUpdate;				//0x30
-	D2RoomCollisionGridStrc* pCollisionGrid;//0x34
-	D2RoomExStrc* pRoomEx;					//0x38
-	D2SeedStrc pSeed;						//0x3C
-	D2DrlgDeleteStrc* pDrlgDelete;			//0x44
-	uint32_t dwFlags;							//0x48
-	D2ClientStrc** ppClients;				//0x4C
-	int32_t nNumClients;						//0x50
-	int32_t nMaxClients;						//0x54
-	int32_t nTileCount;							//0x58
-	int32_t nAllies;							//0x5C // nPlayerUnits in the game
-	int32_t unk0x60;							//0x60
-	uint8_t nCurrentDeathIndex;				//0x64
-	uint8_t pad0x65[3];						//0x65
-	int32_t nLastDeadGUIDs[4];					//0x68
-	D2DrlgActStrc* pAct;					//0x78
-	D2RoomStrc* pRoomNext;					//0x7C
-};
-
-struct D2RoomTileStrc
-{
-	D2RoomExStrc* pRoomEx;					//0x00
-	D2LvlWarpTxt* pLvlWarpTxtRecord;		//0x04
-	BOOL bEnabled;							//0x08
-	D2DrlgTileDataStrc* unk0x0C;			//0x0C
-	D2DrlgTileDataStrc* unk0x10;			//0x10
-	D2RoomTileStrc* pNext;					//0x14
-};
-
-typedef void(__stdcall* AUTOMAPFN)(D2RoomStrc*);
-typedef void(__stdcall* TOWNAUTOMAPFN)(int32_t, int32_t, int32_t, int32_t);
-
-struct D2DrlgStrc
-{
-	D2DrlgLevelStrc* pLevel;				//0x00
-	void* pMempool;							//0x04
-	void* unk0x08;							//0x08
-	D2DrlgActStrc* pAct;					//0x0C
-	uint8_t nAct;								//0x10
-	uint8_t pad0x11[3];						//0x11
-	D2SeedStrc pSeed;						//0x14
-	uint32_t dwStartSeed;						//0x1C
-	uint32_t dwGameLowSeed;					//0x20
-	uint32_t dwFlags;							//0x24
-	D2RoomExStrc pRooms[4];		//0x28
-	D2RoomExStrc* pRoomEx;			//0x3D8
-	uint8_t unk0x3DC;							//0x3DC
-	uint8_t unk0x3DD[3];						//0x3DD
-	int32_t unk0x3E0[2];						//0x3E0
-	D2GameStrc* pGame;						//0x3E8
-	uint8_t nDifficulty;						//0x3EC
-	uint8_t pad0x3ED[3];						//0x3ED
-	AUTOMAPFN pfAutomap;					//0x3F0
-	TOWNAUTOMAPFN pfTownAutomap;			//0x3F4
-	int32_t nStaffTombLevel;					//0x3F8
-	int32_t nBossTombLevel;						//0x3FC
-	D2TileLibraryHashStrc* pTiles[32];		//0x400
-	int32_t bJungleInterlink;					//0x480
-	D2DrlgWarpStrc* pWarp;					//0x484
-};
-
-struct D2DrlgTileDataStrc
-{
-	int32_t nWidth;								//0x00
-	int32_t nHeight;							//0x04
-	int32_t nPosX;								//0x08
-	int32_t nPosY;								//0x0C
-	int32_t unk0x10;							//0x10
-	uint32_t dwFlags;					//0x14
-	D2TileLibraryEntryStrc* pTile;			//0x18
-	int32_t unk0x1C;							//0x1C
-	D2DrlgTileDataStrc* unk0x20;			//0x20
-	int32_t unk0x24;							//0x24
-	uint8_t nRed;								//0x28
-	uint8_t nGreen;							//0x29
-	uint8_t nBlue;								//0x2A
-	uint8_t nIntensity;						//0x2B
-	int32_t unk0x2C;							//0x2C
-};
-
-typedef void(__fastcall* ACTCALLBACKFN)(D2RoomStrc*);
-
-struct D2DrlgActStrc
-{
-	uint8_t nAct;								//0x00
-	uint8_t pad0x01[3];						//0x01
-	D2RoomStrc* pRoom;						//0x04
-	D2DrlgStrc* pDrlg;						//0x08
-	uint32_t dwInitSeed;						//0x0C
-	int32_t nTownId;							//0x10
-	D2DrlgEnvironmentStrc* pEnvironment;	//0x14
-	ACTCALLBACKFN pfnActCallBack;			//0x18
-	BOOL bClient;							//0x1C
-	BOOL unk0x20;							//0x20
-	BOOL bUpdateEx;							//0x24
-	BOOL bUpdate;							//0x28
-	D2DrlgTileDataStrc pTileData;			//0x2C
-	void* pMemPool;							//0x5C
-};
-
-struct D2DrlgAnimTileGridStrc
-{
-	D2DrlgTileDataStrc** ppMapTileData;		//0x00
-	int32_t nFrames;							//0x04
-	int32_t nCurrentFrame;						//0x08
-	int32_t nAnimationSpeed;					//0x0C
-	D2DrlgAnimTileGridStrc* pNext;			//0x10
-};
-
-struct D2DrlgBuildStrc
-{
-	int32_t nPreset;							//0x00
-	int32_t nDivisor;							//0x04
-	int32_t nRand;								//0x08
-	D2DrlgBuildStrc* pNext;					//0x0C
-};
-
-struct D2DrlgDeleteStrc
-{
-	int32_t nUnitType;							//0x00
-	int32_t nUnitGUID;							//0x04
-	D2DrlgDeleteStrc* pNext;				//0x08
-};
-
-struct D2DrlgFileStrc
-{
-	int32_t unk0x00;							//0x00
-	void* pDS1File;							//0x04
-	int32_t unk0x08;							//0x08
-	int32_t nWidth;								//0x0C
-	int32_t nHeight;							//0x10
-	int32_t nWalls;								//0x14
-	int32_t nFloors;							//0x18
-	void* pOrientationLayer[4];				//0x1C
-	void* pWallLayer[4];					//0x2C
-	void* pFloorLayer[2];					//0x3C
-	void* pShadowLayer;						//0x44
-	void* pSubstGroupTags;					//0x48
-	int32_t nSubstGroups;						//0x4C named nClusters in original game
-	void* pSubstGroups;						//0x50
-	D2PresetUnitStrc* pPresetUnit;			//0x54
-	D2DrlgFileStrc* pNext;					//0x58
-};
-
-struct D2DrlgTileInfoStrc
-{
-	int32_t nPosX;								//0x00
-	int32_t nPosY;								//0x04
-	int32_t nTileIndex;							//0x08
-};
-
-struct D2DrlgLevelStrc
-{
-	D2DrlgStrc* pDrlg;						//0x00
-	int32_t nLevelId;							//0x04
-	int32_t nLevelType;							//0x08
-	int32_t nDrlgType;							//0x0C
-	uint32_t dwFlags;							//0x10
-	D2SeedStrc pSeed;						//0x14
-	uint32_t dwInitSeed;						//0x1C
-	union
-	{
-		struct
-		{
-			int32_t nPosX;						//0x20
-			int32_t nPosY;						//0x24
-			int32_t nWidth;						//0x28
-			int32_t nHeight;					//0x2C
-		};
-		D2DrlgCoordStrc pLevelCoords;		//0x20
-	};
-	D2RoomExStrc* pFirstRoomEx;				//0x30
-	int32_t nRooms;								//0x34
-	union
-	{
-		D2DrlgPresetInfoStrc* pPreset;		//0x38
-		D2DrlgOutdoorInfoStrc* pOutdoors;	//0x38
-		D2LvlMazeTxt* pMaze;				//0x38
-
-		//void* pLevelData;						//0x38
-	};
-	D2DrlgMapStrc* pCurrentMap;				//0x3C
-	int32_t nCoordLists;						//0x40
-	D2DrlgTileInfoStrc pTileInfo[32];		//0x44
-	int32_t nTileInfo;							//0x1C4
-	int32_t nRoom_Center_Warp_X[9];				//0x1C8
-	int32_t nRoom_Center_Warp_Y[9];				//0x1EC
-	int32_t nRoomCoords;						//0x210
-	int32_t* pJungleDefs;						//0x214
-	int32_t nJungleDefs;						//0x218
-	D2DrlgBuildStrc* pBuild;				//0x21C
-	BOOL bActive;							//0x220
-	uint32_t dwInactiveFrames;					//0x224
-	int32_t* pPresetMaps;						//0x228
-	D2DrlgLevelStrc* pNextLevel;			//0x22C
-};
-
-struct D2DrlgLevelLinkDataStrc
-{
-	D2SeedStrc pSeed;						//0x00
-	D2DrlgCoordStrc pLevelCoord[15];		//0x08
-	D2DrlgLinkStrc* pLink;					//0xF8
-	union
-	{
-		int32_t nRand[4][15];					//0xFC
-		int32_t nRand2[60];						//0xFC
-	};
-	int32_t nIteration;							//0x1EC
-	int32_t nCurrentLevel;						//0x1F0
-};
-
-struct D2DrlgLinkStrc
-{
-	void* pfLinker;							//0x00
-	int32_t nLevel;								//0x04
-	int32_t nLevelLink;							//0x08
-	int32_t nLevelLinkEx;						//0x0C
-};
-
-struct D2DrlgLinkerParamsStrc
-{
-	int32_t nLinkerLevels[3];					//0x00
-	int32_t nChance[2];							//0x0C
-	int32_t dwFlags;							//0x14
-};
-
-
-struct D2DrlgOrthStrc
-{
-	union
-	{
-		D2RoomExStrc* pRoomEx;				//0x00
-		D2DrlgLevelStrc* pLevel;			//0x00
-	};
-	uint8_t nDirection;						//0x04
-	uint8_t unk0x05[3];						//0x05
-	BOOL bPreset;							//0x08
-	BOOL bInit;								//0x0C
-	D2DrlgCoordStrc* pBox;					//0x10
-	D2DrlgOrthStrc* pNext;					//0x14
-};
-
-
-struct D2DrlgVertexStrc
-{
-	int32_t nPosX;								//0x00
-	int32_t nPosY;								//0x04
-	uint8_t nDirection;						//0x08
-	uint8_t pad0x09[3];						//0x09
-	int32_t dwFlags;							//0x0C
-	D2DrlgVertexStrc* pNext;				//0x10
-};
-
-
-struct D2DrlgPresetInfoStrc
-{
-	D2DrlgMapStrc* pDrlgMap;				//0x00
-	int32_t nDirection;							//0x04
-};
-
-struct D2DrlgRGBStrc
-{
-	uint32_t unk0x00[6];						//0x00
-	uint8_t nRed;								//0x18
-	uint8_t nGreen;							//0x19
-	uint8_t nBlue;								//0x1A
-};
-
-struct D2DrlgRoomTilesStrc
-{
-	D2DrlgTileDataStrc* pWallTiles;			//0x00
-	int32_t nWalls;								//0x04
-	D2DrlgTileDataStrc* pFloorTiles;		//0x08
-	int32_t nFloors;							//0x0C
-	D2DrlgTileDataStrc* pRoofTiles;			//0x10
-	int32_t nRoofs;								//0x14
-};
-
-struct D2DrlgTileGridStrc
-{
-	D2DrlgTileLinkStrc* pMapLinks;			//0x00
-	D2DrlgAnimTileGridStrc* pAnimTiles;		//0x04
-	int32_t nWalls;								//0x08
-	int32_t nFloors;							//0x0C
-	int32_t nShadows;							//0x10
-	D2DrlgRoomTilesStrc pTiles;				//0x14
-};
-
-struct D2DrlgTileLinkStrc
-{
-	BOOL bFloor;							//0x00
-	D2DrlgTileDataStrc* pMapTile;			//0x04
-	D2DrlgTileLinkStrc* pNext;				//0x08
-};
-
-struct D2DrlgUnitsStrc
-{
-	uint32_t unk0x00[9];						//0x00
-	BOOL bDelete;							//0x24
-	BOOL bUpdate;							//0x28
-	uint32_t unk0x2C[12];						//0x2C
-	void* pMemPool;							//0x5C
-};
-
-struct D2DrlgWarpStrc
-{
-	int32_t nLevel;								//0x00
-	int32_t nVis[8];							//0x04
-	int32_t nWarp[8];							//0x24
-	D2DrlgWarpStrc* pNext;					//0x44
-};
-
-struct D2DynamicPathStrc
-{
-	union
-	{
-		struct
-		{
-			uint16_t wOffsetX;					//0x00
-			uint16_t wPosX;						//0x02
-			uint16_t wOffsetY;					//0x04
-			uint16_t wPosY;						//0x06
-		};
-
-		struct
-		{
-			uint32_t dwPrecisionX;				//0x00
-			uint32_t dwPrecisionY;				//0x04
-		};
-	};
-	uint32_t dwTargetX;						//0x08
-	uint32_t dwTargetY;						//0x0C
-	uint16_t xSP1;								//0x10
-	uint16_t ySP1;								//0x12
-	uint16_t xSP2;								//0x14
-	uint16_t ySP2;								//0x16
-	uint16_t xSP3;								//0x18
-	uint16_t ySP3;								//0x1A
-	D2RoomStrc* pRoom;						//0x1C
-	D2RoomStrc* pRoomNext;					//0x20
-	int32_t unk0x24;							//0x24
-	int32_t dwPathPoints;						//0x28
-	void* unk0x2C;							//0x2C
-	D2UnitStrc* pUnit;						//0x30
-	uint32_t dwFlags;							//0x34
-	uint32_t unk0x38;							//0x38
-	uint32_t dwPathType;						//0x3C
-	uint32_t dwPrevPathType;					//0x40
-	uint32_t dwUnitSize;						//0x44
-	uint32_t dwCollisionPattern;				//0x48
-	uint32_t dwCollisionType;					//0x4C
-	uint32_t unk0x50;							//0x50
-	uint16_t unk0x54;							//0x54
-	uint16_t unk0x56;							//0x56
-	D2UnitStrc* pTargetUnit;				//0x58
-	uint32_t dwTargetType;						//0x5C
-	uint32_t dwTargetId;						//0x60
-	uint8_t nDirection;						//0x64
-	uint8_t nNewDirection;						//0x65
-	uint8_t nDiffDirection;					//0x66
-	uint8_t unk0x67;							//0x67
-	uint8_t unk0x68[10];						//0x68
-	int32_t unk0x72;							//0x72
-	int32_t unk0x76;							//0x76
-	char unk0x7A[2];						//0x7A
-	uint32_t dwVelocity;						//0x7C
-	uint32_t unk0x80;							//0x80
-	uint32_t dwVelocity2;						//0x84
-	uint32_t unk0x88[2];						//0x88
-	uint8_t nDist;								//0x90
-	uint8_t nDistMax;							//0x91
-	uint8_t unk0x92;							//0x92
-	uint8_t nStepNum;							//0x93
-	uint8_t nDistance;							//0x94
-	char unk0x95[3];						//0x95
-	//uint32_t dwPathOffset;						//0x98
-	D2PathPointStrc PathPoints[7];			//0x98
-	uint32_t unk0xB8[72];						//0xB8
-	int32_t unk0x1D4;							//0x1D4
-	D2PathPointStrc unk0x1D8[7];			//0x1D8
-	char unk0x1DC[12];						//0x1DC
-};
-
-struct D2EnvironmentCycleStrc
-{
-	int32_t nTicksBegin;			//0x00
-	int32_t nPeriodOfDay;				//0x04
-	uint8_t nRed;					//0x08
-	uint8_t nGreen;				//0x09
-	uint8_t nBlue;					//0x0A
-	uint8_t nIntensity;			//0x0B Unused
-};
 
 struct D2EventListStrc
 {
@@ -1371,35 +881,6 @@ struct D2GfxSettingsStrc
 	BOOL bBlendedShadows;					//0x14
 };
 
-struct D2HirelingInitStrc
-{
-	int32_t nId;								//0x00
-	int32_t nLevel;								//0x04
-	int32_t nHitpoints;							//0x08
-	int32_t nStrength;							//0x0C
-	int32_t nDexterity;							//0x10
-	int32_t nGold;								//0x14
-	int32_t nExperience;						//0x18
-	int32_t nDefense;							//0x1C
-	int32_t nMinDamage;							//0x20
-	int32_t nMaxDamage;							//0x24
-	int32_t nShare;								//0x28
-	int32_t nResist;							//0x2C
-	int32_t nHireDesc;							//0x30
-	int32_t field_34;							//0x34
-	int16_t nFlags;							//0x38
-};
-
-struct D2HoverTextStrc
-{
-	uint32_t dwDisplayTime;					//0x00
-	uint32_t dwExpireTime;						//0x04
-	uint8_t nLangId;							//0x08
-	uint8_t pad0x09[3];						//0x09
-	BOOL bUsed;								//0x0C
-	char szMsg[256];						//0x10
-};
-
 struct D2ImmunityTableStrc
 {
 	int32_t nStat;								//0x00
@@ -1480,35 +961,7 @@ struct D2InteractInfoStrc
 	D2InteractInfoStrc* pNext;				//0x08
 };
 
-struct D2InventoryStrc
-{
-	uint32_t dwSignature;						//0x00
-	void* pMemPool;							//0x04
-	D2UnitStrc* pOwner;						//0x08
-	D2UnitStrc* pFirstItem;					//0x0C
-	D2UnitStrc* pLastItem;					//0x10
-	D2InventoryGridStrc* pGrids;			//0x14
-	int32_t nGridCount;							//0x18
-	uint32_t dwLeftItemGUID;					//0x1C
-	D2UnitStrc* pCursorItem;				//0x20
-	uint32_t dwOwnerId;						//0x24
-	uint32_t dwItemCount;						//0x28
-	D2InventoryNodeStrc* pFirstNode;		//0x2C
-	D2InventoryNodeStrc* pLastNode;			//0x30
-	D2CorpseStrc* pFirstCorpse;				//0x34
-	D2CorpseStrc* pLastCorpse;				//0x38
-	int32_t nCorpseCount;						//0x3C
-};
 
-struct D2InventoryGridStrc
-{
-	D2UnitStrc* pItem;						//0x00
-	D2UnitStrc* pLastItem;					//0x04
-	uint8_t nGridWidth;						//0x08
-	uint8_t nGridHeight;						//0x09
-	uint16_t pad0x0A;							//0x0A
-	D2UnitStrc** ppItems;					//0x0C
-};
 
 //struct D2InventoryInfoStrc
 //{
@@ -1547,57 +1000,6 @@ struct D2ItemCacheStrc
 	int32_t nMagicLevel;						//0x08
 };
 
-struct D2ItemCalcStrc
-{
-	D2UnitStrc* pUnit;						//0x00
-	D2UnitStrc* pItem;						//0x04
-};
-
-struct D2ItemExtraDataStrc
-{
-	D2InventoryStrc* pParentInv;			//0x00
-	D2UnitStrc* pPreviousItem;				//0x04
-	D2UnitStrc* pNextItem;					//0x08
-	char nNodePos;							//0x0C
-	char nNodePosOther;						//0x0D
-	uint16_t unk0x0E;							//0x0E
-	D2UnitStrc* unk0x10;					//0x10
-	D2UnitStrc* unk0x14;					//0x14
-};
-
-struct D2ItemDataStrc
-{
-	uint32_t dwQualityNo;						//0x00
-	D2SeedStrc pSeed;						//0x04
-	uint32_t dwOwnerGUID;						//0x0C
-	uint32_t dwInitSeed;						//0x10
-	uint32_t dwCommandFlags;					//0x14
-	uint32_t dwItemFlags;						//0x18
-	uint32_t dwRealmData[2];					//0x1C
-	uint32_t dwActionStamp;					//0x24
-	int32_t dwFileIndex;						//0x28
-	uint32_t dwItemLevel;						//0x2C
-	uint16_t wItemFormat;						//0x30
-	uint16_t wRarePrefix;						//0x32
-	uint16_t wRareSuffix;						//0x34
-	uint16_t wAutoAffix;						//0x36
-	uint16_t wMagicPrefix[3];					//0x38
-	uint16_t wMagicSuffix[3];					//0x3E
-	uint8_t nBodyLoc;							//0x44					// location on body = xpos of item (only if on body) (spread)
-	uint8_t nInvPage;							//0x45					// 0 = inventory,  3= cube, 4=stash (spread)
-	uint8_t nCellOverlap;						//0x46
-	uint8_t nItemCell;							//0x47
-	uint8_t nEarLvl;							//0x48
-	uint8_t nInvGfxIdx;						//0x49
-	char szPlayerName[16];					//0x4A
-	uint8_t unk0x5A[2];						//0x5A
-	D2ItemExtraDataStrc pExtraData;			//0x5C
-	//D2InventoryStrc* pParentInv;			//0x5C
-	//uint32_t unk0x60;							//0x60
-	//D2UnitStrc* pNextItem;					//0x64
-	//uint8_t nNodePos;							//0x68
-	//uint8_t nNodePosOther;						//0x69
-};
 
 struct D2ItemDescDamageStrc
 {
@@ -1649,19 +1051,6 @@ struct D2ItemDropStrc
 	uint32_t dwFlags2;							//0x80
 };
 
-struct D2ItemSaveStrc
-{
-	int32_t nClassId;							//0x00
-	uint16_t nX;								//0x04
-	uint16_t nY;								//0x06
-	uint8_t nAnimMode;							//0x08
-	uint8_t pad0x09[3];						//0x09
-	uint32_t dwFlags;							//0x0C
-	uint8_t nStorePage;						//0x10
-	uint8_t nBodyloc;							//0x11
-	uint8_t pad0x12[2];						//0x12
-	int32_t nItemFileIndex;						//0x14
-};
 
 struct D2KeyConfigOptionStrc
 {
@@ -1762,111 +1151,6 @@ struct D2MinionListStrc
 	D2MinionListStrc* pNext;				//0x04
 };
 
-struct D2MissileCalcStrc
-{
-	D2UnitStrc* pMissile;					//0x00
-	D2UnitStrc* pOwner;						//0x04
-	int32_t nMissileId;							//0x08
-	int32_t nMissileLevel;						//0x0C
-};
-
-struct D2MissileDamageDataStrc
-{
-	int32_t nFlags;								//0x00
-	int32_t nMinDamage;							//0x04
-	int32_t nMaxDamage;							//0x08
-	int32_t nFireMinDamage;						//0x0C
-	int32_t nFireMaxDamage;						//0x10
-	int32_t nFireLength;						//0x14
-	int32_t nLightMinDamage;					//0x18
-	int32_t nLightMaxDamage;					//0x1C
-	int32_t nMagicMinDamage;					//0x20
-	int32_t nMagicMaxDamage;					//0x24
-	int32_t nColdMinDamage;						//0x28
-	int32_t nColdMaxDamage;						//0x2C
-	int32_t nColdLength;						//0x30
-	int32_t nPoisonMinDamage;					//0x34
-	int32_t nPoisonMaxDamage;					//0x38
-	int32_t nPoisonLength;						//0x3C
-	int32_t nPoisonCount;						//0x40
-	int32_t nLifeDrainMinDamage;				//0x44
-	int32_t nLifeDrainMaxDamage;				//0x48
-	int32_t nManaDrainMinDamage;				//0x4C
-	int32_t nManaDrainMaxDamage;				//0x50
-	int32_t nStaminaDrainMinDamage;				//0x54
-	int32_t nStaminaDrainMaxDamage;				//0x58
-	int32_t nStunLength;						//0x5C
-	int32_t nBurningMin;						//0x60
-	int32_t nBurningMax;						//0x64
-	int32_t nBurnLength;						//0x68
-	int32_t nDemonDamagePercent;				//0x6C
-	int32_t nUndeadDamagePercent;				//0x70
-	int32_t nDamageTargetAC;					//0x74
-	int32_t nDamagePercent;						//0x78
-};
-
-struct D2MissileDataStrc
-{
-	uint32_t unk0x00;							//0x00 - some type of coords, see D2Common.11128, D2Common.11131 - D2Common.11134
-	int16_t nStreamMissile;					//0x04
-	int16_t nStreamRange;						//0x06
-	int16_t nActivateFrame;					//0x08
-	int16_t nSkill;							//0x0A
-	int16_t nLevel;							//0x0C
-	int16_t nTotalFrames;						//0x0E
-	int16_t nCurrentFrame;					//0x10
-	uint16_t unk0x12;							//0x12
-	uint32_t fFlags;							//0x14
-	int32_t nOwnerType;							//0x18
-	uint32_t dwOwnerGUID;						//0x1C
-	int32_t nHomeType;							//0x20
-	uint32_t dwHomeGUID;						//0x24
-	union
-	{
-		struct
-		{
-			int32_t nStatus;					//0x28 - used only by heatseeking projectiles
-			D2PathPointStrc pCoords;		//0x2C - same, these are not coords, they are missile streams etc, see D2Common.11213 & D2Common.11214
-		};
-
-		D2CoordStrc pTargetCoords;			//0x28
-	};
-	D2MissileStreamStrc* pStream;			//0x30
-};
-
-struct D2MissileStrc
-{
-	uint32_t dwFlags;							//0x00
-	D2UnitStrc* pOwner;						//0x04
-	D2UnitStrc* pOrigin;					//0x08
-	D2UnitStrc* pTarget;					//0x0C
-	int32_t nMissile;							//0x10
-	int32_t nX;									//0x14
-	int32_t nY;									//0x18
-	int32_t nTargetX;							//0x1C
-	int32_t nTargetY;							//0x20
-	int32_t nGfxArg;							//0x24
-	int32_t nVelocity;							//0x28
-	int32_t nSkill;								//0x2C
-	int32_t nSkillLevel;						//0x30
-	int32_t nLoops;								//0x34
-	uint32_t unk0x38;							//0x38
-	uint32_t unk0x3C;							//0x3C
-	int32_t nFrame;								//0x40
-	int32_t nActivateFrame;						//0x44
-	int32_t nAttBonus;							//0x48
-	int32_t nRange;								//0x4C
-	int32_t nLightRadius;						//0x50
-	void* pInitFunc;						//0x54
-	void* pInitArgs;						//0x58
-};
-
-struct D2MissileStreamStrc
-{
-	int32_t* unk0x00;							//0x00
-	int32_t unk0x04;							//0x04
-};
-
 struct D2ModeChangeStrc
 {
 	int32_t nMode;								//0x00
@@ -1877,56 +1161,6 @@ struct D2ModeChangeStrc
 	uint8_t unk0x14[4];						//0x14
 	int32_t unk0x18;							//0x18
 	int32_t unk0x1C;							//0x1C
-};
-
-struct D2MonStatsInitStrc
-{
-	int32_t nMinHP;								//0x00
-	int32_t nMaxHP;								//0x04
-	int32_t nAC;								//0x08
-	int32_t nTH;								//0x0C
-	int32_t nExp;								//0x10
-	int32_t nA1MinD;							//0x14
-	int32_t nA1MaxD;							//0x18
-	int32_t nA2MinD;							//0x1C
-	int32_t nA2MaxD;							//0x20
-	int32_t nS1MinD;							//0x24
-	int32_t nS1MaxD;							//0x28
-	int32_t nElMinD;							//0x2C
-	int32_t nElMaxD;							//0x30
-	int32_t nElDur;								//0x34
-};
-
-struct D2MonsterDataStrc
-{
-	D2MonStatsTxt* pMonstatsTxt;			//0x00
-	uint8_t nComponent[16];					//0x04
-	uint16_t wNameSeed;							//0x14
-	uint8_t nTypeFlag;							//0x16
-	uint8_t nLastAnimMode;						//0x17
-	uint32_t dwDurielFlag;						//0x18
-	uint8_t nMonUmod[10];						//0x1C
-	uint16_t wBossHcIdx;						//0x26
-	D2AiControlStrc* pAiControl;			//0x28
-	union									//0x2C
-	{
-		D2AiParamStrc* pAiParam;	//Server pMonster
-		wchar_t* wszMonName;		//Client pMonster
-	};
-	D2MonsterInteractStrc* pMonInteract;	//0x30
-	uint32_t unk0x34[3];						//0x34
-	uint32_t dwNecropetFlag;					//0x40
-	uint32_t unk0x44[3];						//0x44
-	uint32_t pVision;							//0x50
-	uint32_t dwAiState;						//0x54
-	uint32_t dwTxtLevelNo;						//0x58
-	uint16_t wSummonerFlag;						//0x5C
-	uint8_t pad0x5E[2];						//0x5E
-};
-
-struct D2MonsterInteractStrc
-{
-	D2InteractInfoStrc* pInteractInfo;		//0x00
 };
 
 struct D2MonRegDataStrc
@@ -2056,47 +1290,6 @@ struct D2ObeliskPowerUpStrc
 	int32_t nValue;								//0x08
 };
 
-struct D2ObjectRoomCoordStrc
-{
-	D2RoomStrc* pRoom;						//0x00
-	int32_t nX;									//0x04
-	int32_t nY;									//0x08
-	D2ObjectRoomCoordStrc* pNext;			//0x0C
-};
-
-struct D2ObjectRegionStrc
-{
-	uint32_t dwAct;							//0x00
-	uint32_t dwSpawnCount;						//0x04
-	uint8_t unk0x08[8];						//0x08
-	D2ObjectRoomCoordStrc* pObjectRoomCoord;//0x10
-	uint8_t unk0x14[16];						//0x14
-};
-
-struct D2ObjectControlStrc
-{
-	D2SeedStrc pSeed;						//0x00
-	int32_t nShrineTypes[5];					//0x08
-	void* pShrineList[5];					//0x20
-	uint8_t unk0x034[20];						//0x34
-	D2ObjectRegionStrc pReg[1024];			//0x48
-};
-
-struct D2ObjectDataStrc
-{
-	D2ObjectsTxt* pObjectTxt;				//0x00
-	uint8_t InteractType;					//0x04
-	uint8_t nPortalFlags;					//0x05
-	uint16_t unk0x06;						//0x06
-	D2ShrinesTxt* pShrineTxt;				//0x08
-	uint32_t dwOperateGUID;					//0x0C
-	BOOL bPermanent;						//0x10
-	uint32_t __014;							//0x14
-	D2CoordStrc DestRoomCooords;			//0x18
-	D2CoordStrc DestPortalCoords;			//0x20
-	char szOwner[16];						//0x28
-};
-
 typedef void(__fastcall* OBJINITFN)(D2ObjInitFnStrc*);
 
 struct D2ObjInitFnStrc
@@ -2134,94 +1327,6 @@ struct D2PacketTableStrc
 	void* pCallback1;						//0x00
 	int32_t nSize;								//0x04
 	void* pCallback2;						//0x08
-};
-
-struct D2PathInfoStrc
-{
-	D2PathPointStrc pStartCoord;			//0x00
-	D2PathPointStrc field_4;				//0x04
-	D2RoomStrc* pRoom;						//0x08
-	D2RoomStrc* field_C;					//0x0C
-	int32_t field_10;							//0x10
-	uint8_t field_14;							//0x14
-	uint8_t field_15;							//0x15
-	uint16_t field_16;							//0x16
-	int32_t nDistMax;							//0x18
-	uint8_t field_1C;							//0x1C
-	uint8_t field_1D;							//0x1D
-	uint16_t field_1E;							//0x1E
-	int32_t nPathType;							//0x20
-	int32_t nUnitSize;							//0x24
-	int32_t nCollisionPattern;					//0x28
-	int32_t nCollisionType;						//0x2C
-	union
-	{
-		D2DynamicPathStrc* pDynamicPath;	//0x30 - not sure yet
-		D2PathInfoStrc* pNext;				//0x30
-	};
-};
-
-struct D2PlayerCountBonusStrc
-{
-	int32_t nHp;								//0x00
-	int32_t nExperience;						//0x04
-	int32_t nMonsterSkillBonus;					//0x08
-	int32_t nDifficulty;						//0x0C
-	int32_t nPlayerCount;						//0x10
-};
-
-struct D2PlayerDataStrc
-{
-	char szName[16];						//0x00
-	D2BitBufferStrc* pQuestData[3];			//0x10
-	D2WaypointDataStrc* pWaypointData[3];	//0x1C
-	uint32_t unk0x28;							//0x28
-	int32_t nPortalFlags;						//0x2C
-	int32_t unk0x30;							//0x30
-	D2ArenaUnitStrc* pArenaUnit;			//0x34
-	uint32_t unk0x38[4];						//0x38
-	uint32_t dwUniqueId;						//0x48
-	uint32_t dwTradeTick;						//0x4C
-	uint32_t dwTradeState;						//0x50
-	uint32_t unk0x54;							//0x54
-	uint32_t dwAcceptTradeTick;				//0x58
-	D2PlayerTradeStrc* pTrade;				//0x5C
-	uint32_t unk0x60[3];						//0x60
-	uint32_t dwBoughtItemId;					//0x6C
-	int32_t nRightSkillId;						//0x70
-	int32_t nLeftSkillId;						//0x74
-	int32_t nRightSkillFlags;					//0x78
-	int32_t nLeftSkillFlags;					//0x7C
-	int32_t nSwitchRightSkillId;				//0x80
-	int32_t nSwitchLeftSkillId;					//0x84
-	int32_t nSwitchRightSkillFlags;				//0x88
-	int32_t nSwitchLeftSkillFlags;				//0x8C
-	int32_t nWeaponGUID;						//0x90
-	uint32_t unk0x94[2];						//0x94
-	D2ClientStrc* pClient;					//0x9C
-	uint32_t unk0xA0[48];						//0xA0
-	uint32_t dwHostileDelay;					//0x160
-	uint32_t unk0x164;							//0x164
-	uint32_t dwGameFrame;						//0x168
-};
-
-struct D2PlayerTradeStrc
-{
-	int32_t nSaveLength;						//0x00
-	uint8_t* pSaveData;						//0x04
-	int32_t unk0x08[4];
-};
-
-struct D2PresetUnitStrc
-{
-	int32_t nUnitType;							//0x00
-	int32_t nIndex;								//0x04 see D2Common.#11278
-	int32_t nMode;								//0x08
-	int32_t nXpos;								//0x0C
-	int32_t nYpos;								//0x10
-	BOOL bSpawned;							//0x14 
-	D2MapAIStrc* pMapAI;					//0x18
-	D2PresetUnitStrc* pNext;				//0x1C
 };
 
 typedef BOOL(__fastcall* SPELLPREPARE)(D2GameStrc*, D2UnitStrc*, D2UnitStrc*, D2UnitStrc*, int32_t, int32_t, int32_t);
@@ -2490,112 +1595,6 @@ struct D2ResistanceRecordStrc
 	char* szName;							//0x24
 };
 
-struct D2RoomCollisionGridStrc
-{
-	D2DrlgCoordsStrc pRoomCoords;			//0x00
-	uint16_t* pCollisionMask;					//0x20
-};
-
-//struct D2RoomCoordStrc
-//{
-//	int32_t nSubtileX;							//0x00
-//	int32_t nSubtileY;							//0x04
-//	int32_t nSubtileWidth;						//0x08
-//	int32_t nSubtileHeight;						//0x0C
-//	int32_t nTileXPos;							//0x10
-//	int32_t nTileYPos;							//0x14
-//	int32_t nTileWidth;							//0x18
-//	int32_t nTileHeight;						//0x1C
-//};
-
-struct D2RoomCoordListStrc
-{
-	D2DrlgCoordStrc pBox[2];				//0x00
-	BOOL bNode;								//0x20
-	BOOL bRoomActive;						//0x24
-	int32_t nIndex;								//0x28
-	D2RoomCoordListStrc* pNext;				//0x2C
-};
-
-struct D2RosterButtonStrc
-{
-	int32_t unk0x00;							//0x00
-	int32_t nX;									//0x04
-	int32_t nY;									//0x08
-	D2CellFileStrc* pCellFile;				//0x0C
-	int32_t nFrame;								//0x10
-	int32_t nButtonType;						//0x14
-	int32_t nPartyFlags;						//0x18
-	int32_t unk0x1C;							//0x1C
-};
-
-struct D2RosterControlStrc
-{
-	D2RosterButtonStrc pButtons[5];			//0x00
-	char szNameEx[16];						//0xA0
-	uint32_t unk0xB0;							//0xB0
-	uint32_t unk0xB4;							//0xB4
-	uint8_t unk0xB8;							//0xB8
-	uint32_t dwClassId;						//0xB9
-	uint32_t dwLevel;							//0xBD
-	uint32_t dwLevelId;						//0xC1
-	uint32_t dwUnitId;							//0xC5
-	uint32_t unk0xC9;							//0xC9
-	D2RosterControlStrc* pNext;				//0xCD
-};
-
-struct D2RosterCorpseStrc
-{
-	uint32_t dwCorpseId;						//0x00
-	D2RosterCorpseStrc* pPrev;				//0x04
-};
-
-struct D2RosterInfoStrc
-{
-	uint32_t dwUnitId;							//0x00
-	uint32_t dwPartyFlags;						//0x04
-	D2RosterInfoStrc* pNext;				//0x08
-};
-
-struct D2RosterPetStrc
-{
-	int32_t nMonStatsId;						//0x00
-	int32_t nPetTypeId;							//0x04
-	uint32_t dwUnitId;							//0x08
-	uint32_t dwOwnerId;						//0x0C
-	uint32_t unk0x10;							//0x10
-	uint32_t unk0x14;							//0x14
-	uint32_t unk0x18;							//0x18
-	int32_t nLifePercent;						//0x1C
-	BOOL bDrawIcon;							//0x20
-	void* unk0x24;							//0x24
-	uint32_t unk0x28;							//0x28
-	uint32_t unk0x2C;							//0x2C
-	D2RosterPetStrc* pNext;					//0x30
-};
-
-struct D2RosterUnitStrc
-{
-	char szName[16];						//0x00
-	uint32_t dwUnitId;							//0x10
-	uint32_t dwPartyLife;						//0x14
-	uint32_t dwKillCount;						//0x18
-	uint32_t dwClassId;						//0x1C
-	uint16_t wLevel;							//0x20
-	uint16_t wPartyId;							//0x22
-	uint32_t dwLevelId;						//0x24
-	uint32_t dwPosX;							//0x28
-	uint32_t dwPosY;							//0x2C
-	uint32_t dwPartyFlags;						//0x30
-	D2RosterInfoStrc** pRosterInfo;			//0x34
-	D2RosterCorpseStrc* pCorpse;			//0x38
-	uint32_t dwSrcPortalGUID;					//0x3C
-	uint32_t dwDstPortalGUID;					//0x40
-	uint16_t unk0x44;							//0x44
-	uint8_t unk0x46[32];						//0x46
-	char szNameEx[26];						//0x66 - name with clan tag
-	D2RosterUnitStrc* pNext;				//0x80
-};
 
 struct D2SaveHeaderStrc
 {
@@ -2667,160 +1666,10 @@ struct D2ShrineTableStrc
 	int32_t unk0x08;							//0x08
 };
 
-struct D2SkillCalcStrc
-{
-	D2UnitStrc* pUnit;						//0x00
-	int32_t nSkillId;							//0x04
-	int32_t nSkillLevel;						//0x08
-};
-
-struct D2SkillListStrc
-{
-	void* pMemPool;							//0x00
-	D2SkillStrc* pFirstSkill;				//0x04
-	D2SkillStrc* pLeftSkill;				//0x08
-	D2SkillStrc* pRightSkill;				//0x0C
-	D2SkillStrc* pUsedSkill;				//0x10
-	uint32_t __014;							//0x14
-};
-
-//struct D2SkillStrc
-//{
-//	D2SkillsTxt* pSkillsTxt;				//0x00
-//	D2SkillStrc* pNextSkill;				//0x04
-//	uint32_t dwSkillMode;						//0x08
-//	uint32_t unk0x0C[7];						//0x0C
-//	uint32_t dwSkillLevel;						//0x28
-//	uint32_t unk0x2C[2];						//0x2C
-//	uint32_t dwFlags;							//0x34
-//	uint32_t unk0x38;							//0x38
-//};
-struct D2SkillStrc
-{
-
-	D2SkillsTxt* pSkillsTxt;				//0x00
-	D2SkillStrc* pNextSkill;				//0x04
-	uint32_t dwSkillMode;						//0x08
-	uint32_t dwFlags;							//0x0C
-	uint32_t unk0x10[2];						//0x10 - not used?
-
-	union
-	{
-		uint32_t dwTargetInfo;					//0x18
-		uint32_t nXpos;						//0x18
-		uint32_t nPar1;						//0x18
-	};
-	union
-	{
-		uint32_t dwTargetType;					//0x1C
-		uint32_t nYpos;						//0x1C
-		uint32_t nPar2;						//0x1C
-	};
-	union
-	{
-		uint32_t dwTargetGUID;					//0x20
-		uint32_t nPar3;						//0x20
-	};
-
-	uint32_t nPar4;							//0x24
-	int32_t nSkillLevel;						//0x28
-	uint32_t nLevelBonus;						//0x2C
-	int32_t nQuantity;							//0x30
-	int32_t nOwnerGUID;							//0x34 -1 = Native Skill
-	int32_t nCharges;							//0x38
-};
-
-struct D2SkillTreeChartStrc
-{
-	D2CellFileStrc* pCellFile;				//0x00
-	char* szFileName;						//0x04
-	uint32_t unk0x08[6];						//0x08
-	uint16_t unk0x20;							//0x20
-};
-
 struct D2SplittedTextStrc
 {
 	wchar_t* wszLine;						//0x00
 	D2SplittedTextStrc* pNext;				//0x04
-};
-
-struct D2StatStrc
-{
-	union
-	{
-		struct
-		{
-			uint16_t nLayer;					//0x00
-			uint16_t nStat;						//0x02
-		};
-		int32_t nLayer_StatId;					//0x00
-	};
-	int32_t nValue;								//0x04
-};
-
-struct D2StatExStrc
-{
-	D2StatStrc* pStat;						//0x00 An Array[wStatCount]
-	uint16_t nStatCount;						//0x04
-	uint16_t nBitSize;							//0x06
-};
-
-struct D2StatListStrc
-{ 
-	void* pMemPool;							//0x00
-	D2UnitStrc* pUnit;						//0x04
-	uint32_t dwOwnerType;						//0x08
-	uint32_t dwOwnerId;						//0x0C
-	uint32_t dwFlags;							//0x10
-	uint32_t dwStateNo;						//0x14
-	uint32_t dwExpireFrame;					//0x18
-	uint32_t dwSkillNo;						//0x1C
-	uint32_t dwSLvl;							//0x20
-	D2StatExStrc Stats;						//0x24
-	D2StatListStrc* pPrevLink;				//0x2C
-	D2StatListStrc* pNextLink;				//0x30
-	D2StatListStrc* pPrev;					//0x34
-	void* fpStatRemove;						//0x38
-};
-
-struct D2StatListExStrc : public D2StatListStrc
-{
-	//void* pMemPool;							//0x00
-	//D2UnitStrc* pUnit;						//0x04
-	//uint32_t dwOwnerType;						//0x08
-	//uint32_t dwOwnerId;						//0x0C
-	//uint32_t dwFlags;							//0x10
-	//uint32_t dwStateNo;						//0x14
-	//uint32_t dwExpireFrame;					//0x18
-	//uint32_t dwSkillNo;						//0x1C
-	//uint32_t dwSLvl;							//0x20
-	//D2StatExStrc Stats;						//0x24
-	//D2StatListStrc* pPrevLink;				//0x2C
-	//D2StatListStrc* pNextLink;				//0x30
-	//D2StatListStrc* pPrev;					//0x34
-	//void* fpStatRemove;						//0x38
-
-
-	D2StatListStrc* pMyLastList;			//0x3C
-	D2StatListStrc* pMyStats;				//0x40
-	D2UnitStrc* pOwner;						//0x44
-	D2StatExStrc FullStats;					//0x48
-	D2StatExStrc ModStats;					//0x50
-	uint32_t* StatFlags;						//0x58 pt to an  array
-	void* fpCallBack;						//0x5C
-	D2GameStrc* pGame;						//0x60
-};
-
-struct D2StaticPathStrc
-{
-	D2RoomStrc* pRoom;						//0x00
-	int32_t nTargetX;							//0x04
-	int32_t nTargetY;							//0x08
-	int32_t nXPos;								//0x0C
-	int32_t nYPos;								//0x10
-	uint32_t unk0x14[2];						//0x14
-	uint8_t nDirection;						//0x1C
-	uint8_t unk0x1D[3];						//0x1D
 };
 
 struct D2SummonArgStrc
@@ -2881,31 +1730,6 @@ struct D2TCStackStrc
 	int16_t nSuperior;						//0x0C
 	int16_t nNormal;							//0x0E
 	int32_t unk0x10;							//0x10
-};
-
-struct D2TextHeaderStrc
-{
-	void* pMemPool;							//0x00
-	uint16_t nCount;							//0x04
-	uint16_t pad0x06;							//0x06
-	D2TextNodeStrc* pNode;					//0x08
-};
-
-struct D2TextMessageStrc
-{
-	wchar_t* pStringLines[6];				//0x00
-	uint16_t nLines;							//0x18
-	int32_t nColor;								//0x1A
-	uint32_t nEndTick;							//0x1E
-	D2TextMessageStrc* pNext;				//0x22
-};
-
-struct D2TextNodeStrc
-{
-	int16_t nStringId;						//0x00
-	uint16_t pad0x02;							//0x02
-	int32_t nMenu;								//0x04
-	D2TextNodeStrc* pNext;					//0x08
 };
 
 struct D2TileStrc
@@ -3001,74 +1825,7 @@ struct D2TileRecordStrc
 	D2TileRecordStrc* pPrev;				//0x10C
 };
 
-struct D2TimerArgStrc
-{
-	void* pUnk;
-	uint32_t dwUnitId;
-};
 
-struct D2TimerQueueStrc
-{
-	uint32_t dwQueueNo;
-	D2TimerStrc* pActiveTimerList[5][256];
-	D2TimerStrc* pExpiredTimerList[5][256];
-	D2TimerStrc* pTimerListByType[5];
-	void* pTimerList;
-};
-
-struct D2TimerStrc
-{
-	uint16_t wTimerType;						//0x00
-	uint16_t wTimerId;							//0x02
-	uint8_t nQueueNo;							//0x04
-	uint8_t unk0x05[3];						//0x05
-	D2UnitStrc* pUnit;						//0x08
-	uint32_t dwUnitId;							//0x0C
-	uint32_t dwUnitType;						//0x10
-	uint32_t unk0x14;							//0x14
-	D2TimerStrc* pPrevTimer;				//0x18
-	D2TimerStrc* pNextTimer;				//0x1C
-	D2TimerStrc* pPrevTimerOnUnit;			//0x20
-	D2TimerStrc* pNextTimerOnUnit;			//0x24
-	void* pTimerList;						//0x28
-	void* fpTimerFunction;					//0x2C
-};
-
-
-typedef int32_t(__fastcall* UNITFINDTEST)(D2UnitStrc* pUnit, D2UnitFindArgStrc* pUnitFindArg);
-
-struct D2UnitFindArgStrc
-{
-	int32_t nFlags;								//0x00
-	int32_t unk0x04;							//0x04
-	D2UnitStrc* pUnit;						//0x08
-	int32_t nX;									//0x0C
-	int32_t nY;									//0x10
-	int32_t nSize;								//0x14
-	int32_t nMaxArrayEntries;					//0x18
-	int32_t nIndex;								//0x1C
-	D2FieldStrc* pField;					//0x20
-	UNITFINDTEST pfnUnitTest;				//0x24
-	int32_t unk0x28;							//0x28
-	int32_t unk0x2C;							//0x2C
-	int32_t unk0x30;							//0x30
-	int32_t unk0x34;							//0x34
-};
-
-struct D2UnitFindDataStrc
-{
-	void* pMemPool;							//0x00
-	D2UnitStrc** pUnitsArray;				//0x04
-	D2RoomStrc* pRoom;						//0x08
-	UNITFINDTEST pfnUnitTest;				//0x0C
-	D2UnitFindArgStrc* pUnitFindArg;		//0x10
-	int32_t nFlags;								//0x14
-	int32_t nIndex;								//0x18
-	int32_t nMaxArrayEntries;					//0x1C
-	int32_t nX;									//0x20
-	int32_t nY;									//0x24
-	int32_t nSize;								//0x28
-};
 
 struct D2WarpTileStrc
 {
@@ -3078,43 +1835,6 @@ struct D2WarpTileStrc
 	uint32_t __00C;							//0x0C
 	D2LvlWarpTxt* pLvlWarpTxt;				//0x10
 	uint32_t __014;							//0x14
-};
-
-struct D2WaypointActTableStrc
-{
-	bool bTableInitialized;					//0x00
-	int32_t nStartLevelId;						//0x01
-	int32_t nEndLevelId;						//0x05
-	char nWpNo;								//0x09
-	char nLastWpNo;							//0x0A
-};
-
-struct D2WaypointCoordTableStrc
-{
-	int32_t nX;									//0x00
-	int32_t nY;									//0x04
-	int32_t nTextX;								//0x08
-	int32_t nTextY;								//0x0C
-	int32_t nClickX;							//0x10
-	int32_t nClickY;							//0x14
-};
-
-struct D2WaypointDataStrc
-{
-	// Called "history" in the original game
-	uint16_t nFlags[8];							//0x00
-};
-
-struct D2WaypointFlagTableStrc
-{
-	uint16_t nArrayId;							//0x00
-	uint16_t nFlag;								//0x02
-};
-
-struct D2WaypointTableStrc
-{
-	int32_t nLevelId;							//0x00
-	bool bActivated;						//0x04
 };
 
 struct D2WeaponDescTblStrc
